requirements.txt
# Core dependencies
apache-airflow==2.8.0
requests==2.31.0
tenacity==8.2.3
pandas==2.1.4
pyarrow==14.0.2

# Data processing
numpy==1.26.3
fastparquet==2024.2.0

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Logging and monitoring
python-json-logger==2.0.7

# Optional: for local development
jupyter==1.0.0
matplotlib==3.8.2
